db_pipeline/requirements.txt
boto3==1.37.4
botocore==1.37.4
exceptiongroup==1.2.2
iniconfig==2.0.0
jmespath==1.0.1
packaging==24.2
pluggy==1.5.0
psycopg2-binary==2.9.10
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.11.3
six==1.17.0
SQLAlchemy==2.0.38
tomli==2.2.1
typing_extensions==4.12.2
urllib3==1.26.20